Titanic Disaster Analysis - Task 1
This repository contains a Python-based analysis of the Titanic disaster, leveraging popular libraries such as Pandas, NumPy, Matplotlib, Seaborn, and Scikit-learn. The goal of this project was to build a predictive model that can answer questions about the types of passengers who were more likely to survive the sinking of the Titanic. The analysis uses passenger data, including information about their name, age, gender, and socio-economic class, to uncover insights and develop a predictive model.