IMDB Indian Movies Analysis
This project presents a comprehensive analysis of Indian movies on the IMDB platform. The goal was to uncover valuable insights, such as the most frequently appearing actors, the directors who have directed the most movies, the top 10 best-performing actors and directors, most popular movie genres, and so much more. Leveraging the power of Python libraries like Pandas, NumPy, Matplotlib, and Seaborn, I was able to manipulate, analyze, and visualize the data to provide a deep understanding of the Indian movie industry. This analysis serves as a valuable resource for movie enthusiasts, industry professionals, and researchers interested in exploring the rich and diverse world of Indian cinema.